requirements.txt
# Dependencies for the Momentum backend
fastapi>=0.105.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.7
openai>=1.0.0
supabase==2.8.1
jinja2>=3.1.2
pydantic>=1.10.8
httpx>=0.24.1
playwright>=1.47.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
email-validator>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
httpx>=0.24.1
factory-boy>=3.3.0
faker>=19.0.0
pytest-html>=3.2.0
pytest-timeout>=2.1.0
locust>=2.16.0
